% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datecalc.R
\name{datecalc}
\alias{datecalc}
\title{A date difference calculator (Hilmo)}
\usage{
datecalc(df, x)
}
\arguments{
\item{df}{the name of your data you are working with}

\item{x}{The number of maximum number of treatment (hoitojakso_n)}
}
\description{
This function generates necessary columns and calculates the difference between the primary operation and secondary operation
}
\details{
incluusio: Includion variable, "Levy" as primary operation and "Poisto" as secondary.
hoitojakso_n: Number of hospital care periods
oper_seq: sequence of the operations (1 primary, 2 secondary)
levy_pos and poisto_pos: position of the primary (levy) or secondary (poisto) operations in Incluusio column
pvm1:pvmx: Operation dates from pvm list-column
levy_date: Date of primary operation
poisto_date: Date of secondary operation
datediff: Date difference between the primary and secondary operations
}
\keyword{calculator}
\keyword{date}
\keyword{difference}
\keyword{hilmo}
