% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilmo_import_sav.R
\name{Hilmo_import_sav}
\alias{Hilmo_import_sav}
\title{A Hilmo import .sav function}
\usage{
Hilmo_import_sav(filename)
}
\arguments{
\item{filename}{the name of your file you wish to import}
}
\description{
This function allows you to import sav files in more tidy format and it creates ID column from lahtopaiva and tnro
}
\examples{
Hilmo_import_sav()
}
\keyword{hilmo}
\keyword{import}
\keyword{sav}
